Created reusable CarDetaisPage component, and dinamic back to list button
