+++ src/app/(website)/website/[id]/page.tsx
@@ -1,0 +1,9 @@
+import CarDetailsPage from '@/components/features/car-details'
+
+interface PageProps {
+  params: { id: string }
+}
+
+export default async function CarDetailsWebsite(props: PageProps) {
+  return <CarDetailsPage {...props} searchParams={{ source: 'website' }} />
+}
